Allow service instances to use credential store (#8)
The credential-store and credential-admin interfaces were originally
intended to store credentials for service bindings. Some lifecycles also
need to interact with credentials for service instances. Now service
instance and bindings IDs can be used with the credential store.

This pattern was previously used within the test components. Now it is
formally allowed.

Signed-off-by: Scott Andrews <scott@andrews.me>